# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Anekazoo backend.",
        "title": "Anekazoo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/animals": {
            "get": {
                "description": "Get a list of all animals",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Get all animals",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Animal"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new animal to the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Create a new animal",
                "parameters": [
                    {
                        "description": "Animal data",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.CreateAnimal"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Animal created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/animals/{id}": {
            "get": {
                "description": "Get a single animal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Get an animal by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Animal"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an animal's data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Update an existing animal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Animal data",
                        "name": "animal",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Animal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animal updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an animal by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "animals"
                ],
                "summary": "Delete an animal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Animal ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Animal deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Animal": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "legs": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.CreateAnimal": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "legs": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}